Extract validator shared methods to reusable concern
